added FIXME's in Sequent.hs
I'm not yet sure about the code for 'solveSimpleUpK'. It seems to contain a
dead code path and to miss handling PremiseG goals that require up-messages.

Quite some confusion seems to stem from misleading comments on the role of the
different goals. I need to make a pass through the code tomorrow and clean
this up. Best together with Benedikt to check that we understand it the same
way.

git-svn-id: https://svn.inf.ethz.ch/svn/basin/infsec/trunk/projects/FOSSP/tools/tamarin-prover@83545 7e91533c-85bf-0310-9ad9-9686b5989d42