parsing changes: arity 1 notation, signature checks
The parser now allows the implicit tupling notation for function symbols of
arity 1. For example: h(x,y,z) is parsed as h(<x,y,z>).

The parser now checks that the parsed function symbols are actually defined in
the signature. Currently, the default signature combining all built-in
theories is used.

git-svn-id: https://svn.inf.ethz.ch/svn/basin/infsec/trunk/projects/FOSSP/tools/tamarin-prover@83599 7e91533c-85bf-0310-9ad9-9686b5989d42