more precise loop-breakers
We now compute a set of premises of rules that must not be solved
automatically, if looping behaviour is to be excluded. The set is computed
using a greedy algorithm to compute the vertex feedback set of a dataflow
relation computed from the AC-variants of the protocol rules. Changes
required/entailed:

- introduce PremIdx and ConcIdx newtypes to mark premise and conclusion
  indices. This entails replacing NodePrem with (NodeId, PremIdx) and
  analogously NodeCond with (NodeId,ConcIdx)
- the case distinction precomputation has become more extensive, as it uses a
  more precise set of loop-breakers
- fix the Makefile to incorporate the move of CSF'12 case studies
- update all regression tests for CSF'12: all proofs remained the same