fix #5: better handling of built-in special MS rules
Special multiset rewriting rules are now marked using explicit constructors.
Moreover, we use constructors to differentiate between construction and
destruction rules. Note that names of construction rules are prefixed with a
'c' and names of destruction rules are prefixed with a 'd'.

Also adapted intruder variants generation to the change that construction
rules log 'Ded' facts.