Updated regression tests - god are we fast now :-)
@beschmi in case you haven't yet done so, could you also have a look at the
diff and check that it makes sense wrt the changes you made? To me it does
make sense. In most cases, we reduce the time by 50% due to the better
implementation of normalization and fresh name generation.

The STS protocols seem to gain a lot of their time from the better goal
selection strategy. Is that true?