Made the URL to the server stand out more visually.
Maybe it would be even better to report/repeat the URL *after* the server is ready,
because a new user might miss the initial output if there are warnings later on (say, well-formedness warnings).
We do want them to access the URL quickly.

(Maybe even auto-launch using the OS's URL handler?)