Remodeled KAS2 and NAXOS using a new protocol/property refactoring, added DH2.
The idea behind the refactoring is described in 'refactoring-notes.txt'.
It's actually great. And I'm not biased.

The models are much nicer, but the prover takes about three times as
long to prove NAXOS. We should check out what really is causing it
(delayed case distinctions on the matching?) and possibly fix.

Note that the eCK model part is now *identical* to the model for the
KAS2-eCK model, which was the goal we set out to solve.