fix #27: simplify variable indices before display in the GUI
The variable index simplification is performed consistently starting at the
root of the proof tree. This entail a cost linear in the depth of the
currently displayed node.

Note that the simplification is always on. If required, we can add a prefix to
the proof-path that decides whether simplification is desired or not.