fix #83: use the new theory of injective fact instances
The both precomputation and the check for the facts with unique instances 
was wrong. We did not exclude persistent facts in the precomputation and 
we did not check that the terms of the conflicting facts agree.

While we were at it we also changed the definition of unique fact instances
to the slightly more convenient definition of 'injective facts'. See the code for 
the definition. The proper theory will be part of my thesis.

The main reason for switching to injective facts is that it supports 
handle-based crypto better. Just use a fresh handle as the first argument of
a linear fact and do not duplicate it while copying. Then you get a fact symbol
with injective instances.