fix #82: generate proper base and step case for induction
Induction is now implemented as proposed in the thesis. It genereates one
proof case stating that the empty trace might satisfy the property of interest
and another case stating that a trace whose prefix satisfies the property of
interest might satisfy the property of interest.

Along the way this also fixed a bug that 'last-freeness' and 'closedness' were
not checked (properly).