fix #86: turn on eager blackholing
The runtime impact is such that the case precomputation, which runs in
parallel, is computed slightly more efficiently, while long-running sequential
tasks are slightly more expensive. Overall, the time for running all case
studies rises from 1m29s to 1m30s.

This is no win in runtime, but the right way to prepare for future
improvements in parallelism. Thus, I leave the flag turned on.