fix: GUI not working in some cases
We sometimes encountered that the GUI responded with "no such lemma or proof
path". The reason seemed to be empty path segments in the URL. We reckon that
the bug was spurious because only some versions of Yesod handled empty path
segments wrongly. We work around this bug by transforming our paths such that
no empty path segments are generated at all.