Restricted to powers of 3
Changed the rules so that only one single Receive rule takes care of all cases, by modifying the three different Send rules. They are all needed. In this version the 2_double and 2_single rules are commented out, so effectively the protocol is restricted to participants that are powers of 3. (Technically speaking, the sum of 3 powers of 3, when ignoring tree generation, is possible as well.)