Cleandot: Making BNF even more loose to trigger less errors.
Note that this BNF is imprecise in the sense that we don't have explicit
expanders for strings that describe rules, which we should have. This
current version works just fine, but the mismatch between the names in
the BNF and the intended meaning of the strings may come back to us
later.  So at some point this should be cleaned up.