Newlines for the win.
Previously some sections got conflated. This turned out to be cause by a lack of
newline at the end of files, which caused the first line of the next file
to be concatenated onto it. E.g.,

file1
   "This is the last sentence"
file2
   "# This is a section head"

would be parsed as

   "This is the last sentence# This is a section head"

which was not as intended.

To fix, added plenty of newlines with some redundancy.