Show git hash, branch and compile time in version
When using the command line switch --version, tamarin-prover
currently only displays the version number. During development
of tamarin, displaying more detailed information such as the git
hash and branch would be useful to enable faster checking of what
version was used during testing.

This commit includes two new cabal dependencies (gitrev and
template-haskell). The former is not strictly necessary and the
commit could be rewritten to extract the git metadata manually -
but I believe using it improves the supportability of such a minor
feature.