Rule colouring support
Rules can be coloured using a special action __colour (or __color)
which sets the colour of the rule when it is rendered in the dot file,
but otherwise is not displayed.

Usage: include an action __colo(u)r containing a string literal with
the rgb color code in hex. E.g. __colour('FFFFFF') would make the rule
white, while __colour('000000') would make it black. Note that font
colour remains black for now.