Refactor colormap to RGB
HSV was only used for initial generation, before converting to
RGB on use. Since we accept RGB colour codes in actions now, it makes
sense to stick to RGB in the colour map.

Also removed unecessary dependence on Double in some of the colour
functions, when all they need is a RealFrac type