Parse theory arguments and update execution
Having extracted the argument string from a theory, pass them into
the theory loader, parse the arguments from it and update the
arguments that came from the command line (only setting arguments
that weren't set explicitly in the command line).

Due to the way diff, define and quit-on-warning are implemented, it is
probably the case that any theory arguments that affect them are not
currently handled correctly.