Recolouring text based on perceived brightness
Formula is based on the Rec. 709 specification (however, note that Rec.
709 applies to gamma-encoded values with gamma = 2.4. I
haven't bothered to check whether SVG colours are correctly
gamma-adjusted or whether it uses a linear colourspace, and the
primaries on an average monitor might be different than those in the
Rec. 709 spec. Anyway, this should still be 'acceptable' and is easy to
adjust later)