Allow case-studies to be made in parallel
Move the temporary files generated during case-studies make goals to
different locations dependant on the case-study path. This allows
multiple case-studies to be verified in parallel (i.e. with the -j
flag on make).

Also use printf to allow the newlines to be output, rather than
echo "\n" which prints a literal \n.