Add installation instructions for Nix/NixOS users
As of a recent version of Nixpkgs, `tamarin-prover` is a maintained, upstream package, along with Maude/SAPIC configured for it, out of the box. We're using a recent snapshot version of tamarin 1.3.0 from the `develop` branch, built with GHC 8.2.2. I've tested the TLS handshake examples with this build which seems to work fine.

This simply adds manual instructions that Nix/NixOS users will be familiar with.