diff mode: conclude to attack as soon as the system is trivial
If a system is not mirrorable and is trivial (for example contains
only !KU(x) goals with x unconstrained) then we know that if we solve
all these goals with public values then it will be an attack.
Unfortunately, this is only one of the cases left to explore, so waiting
for the proof to reach this attack is expensive. Instead we detect this
situation early and immediately conclude to an attack.