SAPIC: apply let substitutions in message patterns.
Note that the formal semantics consider input to be binding. It is very
convenient to use pattern matching at this place though, so we should
introduce a "strict" mode in the future, that adheres to the semantics
formally proven secure, and give out a warning if there is a possible
confusion.