Add SAPiC locking restrictions based on locking / unlocking positions
This change fixes the issue, where a locking restriction has been
added without a corresponding unlock action being present.
This caused a wellformedness warning.

If a lock and the corresponding unlock action are present, the full locking
restriction is added. If only a lock is present, a simplified restriction
is used.