Export and import rule variants (#388)
This PR adds the possibility to export and re-import variants of rules and left or right instances of rules in diff-mode. This is sometimes necessary to be able to export and re-import theories with auto-generated sources lemmas.
When re-importing variants, Tamarin checks the imported variants against recomputed variants, and shows a wellformedness warning if they do not match. The same applies to left&right rule instances in diff-mode.

Commits:
* import and export rule variants and left and right rules explictly

* update regression files to fit new output order

* import and export rule variants and left/right rules