Refactor dot & JSON generation.
Previsouly the dot & JSON generation was done using a constraint system
object. Since both are essentially serializations of a graph structure
this led to some duplicated work. We introduce a new graph data
structure that can be generated from a constraint system and which
presents an abstract interface to both the dot & JSON generation
functions to get the system's nodes, edges etc.

All options set via the GUI (except the compact node toggle which is
only used for dot) change the representation of the graph so that they
are respected by both dot & JSON.

There are still some tasks left to do like the generation of clusters.