Quick fix to preserve options in a proof.
The previous behavior was that if you change any options away from the
default (e.g. turning off abbreviations or setting a simplification
level other than 2) the UI would send a request to the server to render
the current constraint graph under the new settings and update the
current page.
However, when doing a proof step (e.g. solve X) the server sends a
complete new HTML page describing the updated proof state, which then
triggers a new request to render the constraint graph but using the
default settings, even though the settings are still changed in the UI.

In this commit we add a quick hack to pass the image options via GET
parameters when triggering the full page reload. The server will then
generate an HTML page that includes an image URL with these parameters,
so that the new image is generated correctly.
It would also possible to do everything client-side, i.e. when receiving
the HTML page we rewrite the image URL locally. But this results in a
higher latency, as the browser first has to execute JS code before
requesting the image.