Add more comfortable debug tracing functions.
We add the function etraceLn (and etraceLnM) which is like
Debug.trace (and traceM) but only outputs the trace if the supplied key
is contained in the "DEBUG_TRACE" environment variable.
This makes print debugging more comfortable since it's easy to add lots
of print statements and selectively enable them.