Implement dumping found traces to dot/JSON.
In batch mode we add two command line options "--output-json" &
"--output-dot" to output any found traces (either for an exists-trace
lemma, or if an attack is found for an all-traces lemma) to a dot/json
file.
All constraint systems are output into the same file. For dot this means
simply concatenating the graphs. Calling the dot command line tool with
the "-O" option will then generate individual pictures for each
contained graph. For JSON the existing schema already allows for
multiple graphs in one file so we use this.
What is missing is controlling the rendering/graph simplifications
options for the constraint systems. At the moment is always uses the
default options.