Improved abbreviation generation.
We compute the weight only as screen size * occurrences now.
Additionally we keep track of the existing abbreviations so that we
1. subtract a number of occurrences. If term B appears n times as a
subterm of A and we make an abbreviation of A, then we tentatively subtract n
occurrences from B.
2. compute the screen size more accurrately. If term B appears as a
subterm of A and we make an abbreviation of B, then when we next compute
the screen size of A we use the of B to get a more accurrate
measurement.
This does not handle future abbreviations so it's not perfectly precise
but seems to be good enough and it's easier than computing a global
maximum of saved screen space.