Fix error when generating dot for collapsed nodes.
The dot generation uses the DotState structure to pass mappings from
constraint system nodes to graphviz nodes from the node generation pass
to the edge generation pass (since edges are defined on constraint
system nodes but must reference graphviz nodes).
Previously we only filled the dsNodes mapping but when generating
collapsed nodes we actually need to update all mappings. To this end we
define a function to run the dot generation for all collapsed nodes,
throw away the dot result and only keep the mappings to update
the DotState in the actual dot computation.