Warning for not subterm convergency (#663)
Add a function testing the sub term Convergency of the equations of the theory and a warning message for equations that are not subterm convergent. The warning indicates that these equations might enter into cycles or fail to reach a conclusion. This addition aims to help users identify and address potential issues in their equations.

* warning non optimal

* correct warning

* cleaning up and optimizations

* reformulation warning

* fix bug

* functional version before cleaning up

* small test and cleaning

* Added manual section

* removed unused imports, added documentation for  attribute

* export convergent

* added convergent to failing tests

* addes space

* added convergent attribute to grammar

---------

Co-authored-by: Jannik Dreier <research@jannikdreier.net>