Attempt at fixing deployment error for manual.
It seems that github is currently updating `ubuntu-latest` from 22.04 to 24.04.

Currently, manual deployment fails. Build logs indicated that puenc-greek.def is missing. It seems this could be a missing `greek-fontenc` but it is unclear how this could be a new behavior.

I can't find any such issues online with fresh 24.04 installs. However, I can find users reporting a long list of missing latex packages when updating from 22.04 to 24.04 -- yet this doesn't seem like something github should be doing. The solution that might work is installing `texlive-full`, but this seems extremely inefficient.

This is my optimistic attempt at fixing the issue if it only concerns very few packages. Fingers crossed!